+++ src/refactoring/domains/admin/components/index.ts
@@ -1,0 +1,2 @@
+export * from './coupon/index';
+export * from './product/index';
