+++ src/refactoring/domains/admin/components/product/index.ts
@@ -1,0 +1,1 @@
+export * from './ProductManager';
