+++ src/refactoring/domains/cart/components/coupon/CouponSelector.tsx
@@ -1,4 +1,4 @@
-import type { Coupon } from '../../../../types';
+import type { Coupon } from '../../../../../types';
 
 interface CouponSelectorProps {
   coupons: Coupon[];
