+++ src/refactoring/domains/cart/components/models/getMaxDiscount.ts
@@ -1,0 +1,3 @@
+export const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
