+++ src/refactoring/domains/cart/hooks/useCart.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { calculateCartTotal, updateCartItemQuantity } from '../../../models/cart';
+import { calculateCartTotal, getRemainingStock, updateCartItemQuantity } from '../../../models/cart';
 
 import type { CartItem, Coupon, Product } from '../../../../types';
 
@@ -8,13 +8,8 @@ export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
-  const getRemainingStock = (product: Product) => {
-    const cartItem = cart.find((item) => item.product.id === product.id);
-    return product.stock - (cartItem?.quantity || 0);
-  };
-
   const addToCart = (product: Product) => {
-    const remainingStock = getRemainingStock(product);
+    const remainingStock = getRemainingStock(product, cart);
     if (remainingStock <= 0) return;
 
     setCart((prevCart) => {
